backend/data_collection/serializers.py
from rest_framework import serializers

from .models import Location, ProcessedData, RawData


class RawDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = RawData
        fields = '__all__'
        read_only_fields = ('collected_at',)


class ProcessedDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProcessedData
        fields = '__all__'
        read_only_fields = ('processed_at',)


class LocationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Location
        fields = '__all__'
        read_only_fields = ('created_at',)