backend/pollution_prediction/urls.py
from django.contrib import admin
from django.http import HttpResponse
from django.urls import include, path
from rest_framework_simplejwt.views import (TokenObtainPairView,
                                            TokenRefreshView)

urlpatterns = [
    path('admin/', admin.site.urls),
    # API routes
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/users/', include('users.urls')),
    path('api/data/', include('data_collection.urls')),
    path('api/predictions/', include('predictions.urls')),
path('', lambda request: HttpResponse("Django via Nginx fonctionne ✅")),
]