backend/requirements.txt
# Django and REST framework
Django==4.2.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
dj-database-url==2.1.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Kafka
confluent-kafka==2.3.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
xgboost>=1.7.0

# API clients
requests==2.31.0
# Logging
loguru==0.7.2

# Email
resend==0.7.2

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Development
black==23.12.0
isort==5.13.2
flake8==6.1.0

# Production
gunicorn==21.2.0